[ci] Adding post build script (#1295)
After discussion and revert from #1291 , we decided to create a post
build script to help prevent indirection for developers.

This script combines previous upload scripts into one script, with
options for CI and upload.
